+++ vector/src/main/java/im/vector/app/features/home/room/list/RoomListViewModel.kt
@@ -33,6 +33,7 @@ import org.matrix.android.sdk.api.session.Session
 import org.matrix.android.sdk.api.session.room.model.Membership
 import org.matrix.android.sdk.api.session.room.model.RoomSummary
 import org.matrix.android.sdk.api.session.room.model.tag.RoomTag
+import org.matrix.android.sdk.api.session.room.state.isPublic
 import org.matrix.android.sdk.rx.rx
 import timber.log.Timber
 import java.lang.Exception
@@ -76,6 +77,10 @@ class RoomListViewModel @Inject constructor(initialState: RoomListViewState,
             is RoomListAction.ChangeRoomNotificationState -> handleChangeNotificationMode(action)
             is RoomListAction.ToggleTag                   -> handleToggleTag(action)
         }.exhaustive
+    }
+
+    fun isPublicRoom(roomId: String): Boolean {
+        return session.getRoom(roomId)?.isPublic().orFalse()
     }
 
     // PRIVATE METHODS *****************************************************************************