+++ CHANGES.md
@@ -5,15 +5,18 @@ Features ✨:
  - Enable url previews for notices (#2562)
 
 Improvements 🙌:
-  - Add System theme option and set as default (#904) (#2387)
+ - Add System theme option and set as default (#904, #2387)
+ - Warn user when he is leaving a not public room (#1460)
 
 Bugfix 🐛:
  - Unspecced msgType field in m.sticker (#2580)
  - Wait for all room members to be known before sending a message to a e2e room (#2518)
  - Url previews sometimes attached to wrong message (#2561)
+ - Room Topic not displayed correctly after visiting a link (#2551)
  - Hiding membership events works the exact opposite (#2603)
  - Tapping drawer having more than 1 room in notifications gives "malformed link" error (#2605)
  - Sent image not displayed when opened immediately after sending (#409)
+ - Initial sync is not retried correctly when there is some network error. (#2632)
 
 Translations 🗣:
  -
@@ -22,7 +25,7 @@ SDK API changes ⚠️:
  -
 
 Build 🧱:
- -
+ - Remove dependency to org.greenrobot.eventbus library
 
 Test:
  -